Notes fix

+++ src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/NoteController.java
@@ -6,6 +6,8 @@ import com.udacity.jwdnd.course1.cloudstorage.services.CredentialService;
 import com.udacity.jwdnd.course1.cloudstorage.services.FileService;
 import com.udacity.jwdnd.course1.cloudstorage.services.NoteService;
 import com.udacity.jwdnd.course1.cloudstorage.services.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -22,6 +24,8 @@ public class NoteController {
   private FileService fileService;
   private CredentialService credentialService;
 
+  private Logger log = LoggerFactory.getLogger(NoteController.class);
+
   public NoteController(UserService userService, NoteService noteService, FileService fileService, CredentialService credentialService) {
     this.userService = userService;
     this.noteService = noteService;
@@ -34,8 +38,9 @@ public class NoteController {
     String noteError = null;
     String noteUploadError = null;
     User curr = userService.getUser(authentication.getName());
-    if(note.getNoteId() != null) {
+    if(note.getNoteId() == null) {
       try {
+        note.setUserid(curr.getUserid());
         noteService.createNote(note);
         model.addAttribute("noteUploadSuccess", "Note created");
 
@@ -65,16 +70,20 @@ public class NoteController {
   public String deleteNote(Authentication authentication, @ModelAttribute Note note, Model model) throws IOException {
     String deleteError = null;
     User curr = userService.getUser(authentication.getName());
-    try {
-      noteService.deleteNote(note);
-      model.addAttribute("deleteSuccess", "Note Deleted");
-    } catch (Exception e) {
-      deleteError = e.toString();
-      model.addAttribute("deleteError", deleteError);
+    note = noteService.getNote(note.getNoteId());
+    if(note != null){
+      try {
+        noteService.deleteNote(note);
+        model.addAttribute("deleteSuccess", "Note Deleted");
+      } catch (Exception e) {
+        deleteError = e.toString();
+        model.addAttribute("deleteError", deleteError);
+      }
+      model.addAttribute("files", fileService.getFiles(curr));
+      model.addAttribute("notes", noteService.getNotes(curr));
+      model.addAttribute("credentials", credentialService.getAllCredentials(curr));
     }
-    model.addAttribute("files", fileService.getFiles(curr));
-    model.addAttribute("notes", noteService.getNotes(curr));
-    model.addAttribute("credentials", credentialService.getAllCredentials(curr));
+   
     return "home";
   }
 }
